+++ contracts/contracts/Events.sol
@@ -13,19 +13,10 @@ interface Events {
     event BlockVerification(uint32 indexed blockNumber);
 
     /// @notice Event emitted when user send a transaction to withdraw her funds from onchain balance
-    event OnchainWithdrawal(
-        address indexed owner,
-        uint16 indexed tokenId,
-        uint128 amount
-    );
+    event OnchainWithdrawal(address indexed owner, uint16 indexed tokenId, uint128 amount);
 
     /// @notice Event emitted when user send a transaction to deposit her funds
-    event OnchainDeposit(
-        address indexed sender,
-        uint16 indexed tokenId,
-        uint128 amount,
-        address indexed owner
-    );
+    event OnchainDeposit(address indexed sender, uint16 indexed tokenId, uint128 amount, address indexed owner);
 
     /// @notice Event emitted when user sends a authentication fact (e.g. pub-key hash)
     event FactAuth(address indexed sender, uint32 nonce, bytes fact);
@@ -42,7 +33,7 @@ interface Events {
         uint64 serialId,
         Operations.OpType opType,
         bytes pubData,
-        uint256 expirationBlock
+        uint expirationBlock
     );
 
     /// @notice Deposit committed event.
@@ -69,31 +60,28 @@ interface Events {
 
     /// @notice Pending withdrawals index range that were executed in the completeWithdrawals operation.
     /// NOTE: processed indexes in the queue map are [queueStartIndex, queueEndIndex)
-    event PendingWithdrawalsComplete(
-        uint32 queueStartIndex,
-        uint32 queueEndIndex
-    );
+    event PendingWithdrawalsComplete(uint32 queueStartIndex, uint32 queueEndIndex);
 }
 
 /// @title Upgrade events
 /// @author Matter Labs
 interface UpgradeEvents {
     /// @notice Event emitted when new upgradeable contract is added to upgrade gatekeeper's list of managed contracts
-    event NewUpgradable(uint256 indexed versionId, address indexed upgradeable);
+    event NewUpgradable(uint indexed versionId, address indexed upgradeable);
 
     /// @notice Upgrade mode enter event
     event NoticePeriodStart(
-        uint256 indexed versionId,
+        uint indexed versionId,
         address[] newTargets,
-        uint256 noticePeriod // notice period (in seconds)
+        uint noticePeriod // notice period (in seconds)
     );
 
     /// @notice Upgrade mode cancel event
-    event UpgradeCancel(uint256 indexed versionId);
+    event UpgradeCancel(uint indexed versionId);
 
     /// @notice Upgrade mode preparation status event
-    event PreparationStart(uint256 indexed versionId);
+    event PreparationStart(uint indexed versionId);
 
     /// @notice Upgrade mode complete event
-    event UpgradeComplete(uint256 indexed versionId, address[] newTargets);
+    event UpgradeComplete(uint indexed versionId, address[] newTargets);
 }
