+++ contracts/contracts/Operations.sol
@@ -50,17 +50,12 @@ library Operations {
         address owner;
     }
 
-    uint256 public constant PACKED_DEPOSIT_PUBDATA_BYTES =
-        ACCOUNT_ID_BYTES + TOKEN_BYTES + AMOUNT_BYTES + ADDRESS_BYTES;
+    uint public constant PACKED_DEPOSIT_PUBDATA_BYTES = ACCOUNT_ID_BYTES + TOKEN_BYTES + AMOUNT_BYTES + ADDRESS_BYTES;
 
     /// Deserialize deposit pubdata
-    function readDepositPubdata(bytes memory _data)
-        internal
-        pure
-        returns (Deposit memory parsed)
-    {
+    function readDepositPubdata(bytes memory _data) internal pure returns (Deposit memory parsed) {
         // NOTE: there is no check that variable sizes are same as constants (i.e. TOKEN_BYTES), fix if possible.
-        uint256 offset = 0;
+        uint offset = 0;
         (offset, parsed.accountId) = Bytes.readUInt32(_data, offset); // accountId
         (offset, parsed.tokenId) = Bytes.readUInt16(_data, offset); // tokenId
         (offset, parsed.amount) = Bytes.readUInt128(_data, offset); // amount
@@ -70,11 +65,7 @@ library Operations {
     }
 
     /// Serialize deposit pubdata
-    function writeDepositPubdata(Deposit memory op)
-        internal
-        pure
-        returns (bytes memory buf)
-    {
+    function writeDepositPubdata(Deposit memory op) internal pure returns (bytes memory buf) {
         buf = abi.encodePacked(
             bytes4(0), // accountId (ignored) (update when ACCOUNT_ID_BYTES is changed)
             op.tokenId, // tokenId
@@ -84,24 +75,10 @@ library Operations {
     }
 
     /// @notice Check that deposit pubdata from request and block matches
-    function depositPubdataMatch(bytes memory _lhs, bytes memory _rhs)
-        internal
-        pure
-        returns (bool)
-    {
+    function depositPubdataMatch(bytes memory _lhs, bytes memory _rhs) internal pure returns (bool) {
         // We must ignore `accountId` because it is present in block pubdata but not in priority queue
-        bytes memory lhs_trimmed =
-            Bytes.slice(
-                _lhs,
-                ACCOUNT_ID_BYTES,
-                PACKED_DEPOSIT_PUBDATA_BYTES - ACCOUNT_ID_BYTES
-            );
-        bytes memory rhs_trimmed =
-            Bytes.slice(
-                _rhs,
-                ACCOUNT_ID_BYTES,
-                PACKED_DEPOSIT_PUBDATA_BYTES - ACCOUNT_ID_BYTES
-            );
+        bytes memory lhs_trimmed = Bytes.slice(_lhs, ACCOUNT_ID_BYTES, PACKED_DEPOSIT_PUBDATA_BYTES - ACCOUNT_ID_BYTES);
+        bytes memory rhs_trimmed = Bytes.slice(_rhs, ACCOUNT_ID_BYTES, PACKED_DEPOSIT_PUBDATA_BYTES - ACCOUNT_ID_BYTES);
         return keccak256(lhs_trimmed) == keccak256(rhs_trimmed);
     }
 
@@ -114,16 +91,11 @@ library Operations {
         uint128 amount;
     }
 
-    uint256 public constant PACKED_FULL_EXIT_PUBDATA_BYTES =
-        ACCOUNT_ID_BYTES + ADDRESS_BYTES + TOKEN_BYTES + AMOUNT_BYTES;
+    uint public constant PACKED_FULL_EXIT_PUBDATA_BYTES = ACCOUNT_ID_BYTES + ADDRESS_BYTES + TOKEN_BYTES + AMOUNT_BYTES;
 
-    function readFullExitPubdata(bytes memory _data)
-        internal
-        pure
-        returns (FullExit memory parsed)
-    {
+    function readFullExitPubdata(bytes memory _data) internal pure returns (FullExit memory parsed) {
         // NOTE: there is no check that variable sizes are same as constants (i.e. TOKEN_BYTES), fix if possible.
-        uint256 offset = 0;
+        uint offset = 0;
         (offset, parsed.accountId) = Bytes.readUInt32(_data, offset); // accountId
         (offset, parsed.owner) = Bytes.readAddress(_data, offset); // owner
         (offset, parsed.tokenId) = Bytes.readUInt16(_data, offset); // tokenId
@@ -132,11 +104,7 @@ library Operations {
         require(offset == PACKED_FULL_EXIT_PUBDATA_BYTES, "rfp10"); // reading invalid full exit pubdata size
     }
 
-    function writeFullExitPubdata(FullExit memory op)
-        internal
-        pure
-        returns (bytes memory buf)
-    {
+    function writeFullExitPubdata(FullExit memory op) internal pure returns (bytes memory buf) {
         buf = abi.encodePacked(
             op.accountId, // accountId
             op.owner, // owner
@@ -146,16 +114,10 @@ library Operations {
     }
 
     /// @notice Check that full exit pubdata from request and block matches
-    function fullExitPubdataMatch(bytes memory _lhs, bytes memory _rhs)
-        internal
-        pure
-        returns (bool)
-    {
+    function fullExitPubdataMatch(bytes memory _lhs, bytes memory _rhs) internal pure returns (bool) {
         // `amount` is ignored because it is present in block pubdata but not in priority queue
-        uint256 lhs =
-            Bytes.trim(_lhs, PACKED_FULL_EXIT_PUBDATA_BYTES - AMOUNT_BYTES);
-        uint256 rhs =
-            Bytes.trim(_rhs, PACKED_FULL_EXIT_PUBDATA_BYTES - AMOUNT_BYTES);
+        uint lhs = Bytes.trim(_lhs, PACKED_FULL_EXIT_PUBDATA_BYTES - AMOUNT_BYTES);
+        uint rhs = Bytes.trim(_rhs, PACKED_FULL_EXIT_PUBDATA_BYTES - AMOUNT_BYTES);
         return lhs == rhs;
     }
 
@@ -169,24 +131,20 @@ library Operations {
         address owner;
     }
 
-    function readPartialExitPubdata(bytes memory _data, uint256 _offset)
+    function readPartialExitPubdata(bytes memory _data, uint _offset)
         internal
         pure
         returns (PartialExit memory parsed)
     {
         // NOTE: there is no check that variable sizes are same as constants (i.e. TOKEN_BYTES), fix if possible.
-        uint256 offset = _offset + ACCOUNT_ID_BYTES; // accountId (ignored)
+        uint offset = _offset + ACCOUNT_ID_BYTES; // accountId (ignored)
         (offset, parsed.tokenId) = Bytes.readUInt16(_data, offset); // tokenId
         (offset, parsed.amount) = Bytes.readUInt128(_data, offset); // amount
         offset += FEE_BYTES; // fee (ignored)
         (offset, parsed.owner) = Bytes.readAddress(_data, offset); // owner
     }
 
-    function writePartialExitPubdata(PartialExit memory op)
-        internal
-        pure
-        returns (bytes memory buf)
-    {
+    function writePartialExitPubdata(PartialExit memory op) internal pure returns (bytes memory buf) {
         buf = abi.encodePacked(
             bytes4(0), // accountId (ignored) (update when ACCOUNT_ID_BYTES is changed)
             op.tokenId, // tokenId
@@ -207,24 +165,16 @@ library Operations {
         address target;
     }
 
-    function readForcedExitPubdata(bytes memory _data, uint256 _offset)
-        internal
-        pure
-        returns (ForcedExit memory parsed)
-    {
+    function readForcedExitPubdata(bytes memory _data, uint _offset) internal pure returns (ForcedExit memory parsed) {
         // NOTE: there is no check that variable sizes are same as constants (i.e. TOKEN_BYTES), fix if possible.
-        uint256 offset = _offset + ACCOUNT_ID_BYTES * 2; // initiatorAccountId + targetAccountId (ignored)
+        uint offset = _offset + ACCOUNT_ID_BYTES * 2; // initiatorAccountId + targetAccountId (ignored)
         (offset, parsed.tokenId) = Bytes.readUInt16(_data, offset); // tokenId
         (offset, parsed.amount) = Bytes.readUInt128(_data, offset); // amount
         offset += FEE_BYTES; // fee (ignored)
         (offset, parsed.target) = Bytes.readAddress(_data, offset); // target
     }
 
-    function writeForcedExitPubdata(ForcedExit memory op)
-        internal
-        pure
-        returns (bytes memory buf)
-    {
+    function writeForcedExitPubdata(ForcedExit memory op) internal pure returns (bytes memory buf) {
         buf = abi.encodePacked(
             bytes4(0), // initiatorAccountId (ignored) (update when ACCOUNT_ID_BYTES is changed)
             bytes4(0), // targetAccountId (ignored) (update when ACCOUNT_ID_BYTES is changed)
@@ -246,12 +196,12 @@ library Operations {
         //uint16 fee; -- present in pubdata, ignored at serialization
     }
 
-    function readChangePubKeyPubdata(bytes memory _data, uint256 _offset)
+    function readChangePubKeyPubdata(bytes memory _data, uint _offset)
         internal
         pure
         returns (ChangePubKey memory parsed)
     {
-        uint256 offset = _offset;
+        uint offset = _offset;
         (offset, parsed.accountId) = Bytes.readUInt32(_data, offset); // accountId
         (offset, parsed.pubKeyHash) = Bytes.readBytes20(_data, offset); // pubKeyHash
         (offset, parsed.owner) = Bytes.readAddress(_data, offset); // owner
@@ -260,7 +210,7 @@ library Operations {
 
     // Withdrawal data process
 
-    function readWithdrawalData(bytes memory _data, uint256 _offset)
+    function readWithdrawalData(bytes memory _data, uint _offset)
         internal
         pure
         returns (
@@ -270,7 +220,7 @@ library Operations {
             uint128 _amount
         )
     {
-        uint256 offset = _offset;
+        uint offset = _offset;
         (offset, _addToPendingWithdrawalsQueue) = Bytes.readBool(_data, offset);
         (offset, _to) = Bytes.readAddress(_data, offset);
         (offset, _tokenId) = Bytes.readUInt16(_data, offset);
